refactored - workspace handling extracted to module
